Refactor files/packages structure.
- Move project.go to resourcemanager
- Change converters/google/resources to services
