+++ cai2hcl/common/utils.go
