+++ cai2hcl/converter_map.go
@@ -1,22 +1,22 @@
-package cai2hcl
+package convert
 
 import (
-	"github.com/GoogleCloudPlatform/terraform-google-conversion/v2/cai2hcl/converters/google/common"
-	"github.com/GoogleCloudPlatform/terraform-google-conversion/v2/cai2hcl/converters/google/resources"
-	"github.com/GoogleCloudPlatform/terraform-google-conversion/v2/cai2hcl/converters/google/resources/compute"
+	"github.com/GoogleCloudPlatform/terraform-google-conversion/v2/cai2hcl/common"
+	"github.com/GoogleCloudPlatform/terraform-google-conversion/v2/cai2hcl/services/compute"
+	"github.com/GoogleCloudPlatform/terraform-google-conversion/v2/cai2hcl/services/resourcemanager"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	tpg "github.com/hashicorp/terraform-provider-google-beta/google-beta"
 )
 
 var ConverterNames = map[string]string{
-	resources.ProjectAssetType:             "google_project",
-	resources.ProjectBillingAssetType:      "google_project",
-	compute.ComputeInstanceAssetType:       "google_compute_instance",
-	compute.ComputeForwardingRuleAssetType: "google_compute_forwarding_rule",
+	resourcemanager.ProjectAssetType:        "google_project",
+	resourcemanager.ProjectBillingAssetType: "google_project",
+	compute.ComputeInstanceAssetType:        "google_compute_instance",
+	compute.ComputeForwardingRuleAssetType:  "google_compute_forwarding_rule",
 }
 
 var converterFactories = map[string]func(name string, schema map[string]*schema.Schema) common.Converter{
-	"google_project":                 resources.NewProjectConverter,
+	"google_project":                 resourcemanager.NewProjectConverter,
 	"google_compute_instance":        compute.NewComputeInstanceConverter,
 	"google_compute_forwarding_rule": compute.NewComputeForwardingRuleConverter,
 }
