+++ cai2hcl/services/compute/compute_forwarding_rule.go
@@ -3,7 +3,7 @@ package compute
 import (
 	"fmt"
 
-	"github.com/GoogleCloudPlatform/terraform-google-conversion/v2/cai2hcl/converters/google/common"
+	"github.com/GoogleCloudPlatform/terraform-google-conversion/v2/cai2hcl/common"
 	"github.com/GoogleCloudPlatform/terraform-google-conversion/v2/caiasset"
 
 	tfschema "github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
