+++ cai2hcl/services/compute/compute_instance.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/GoogleCloudPlatform/terraform-google-conversion/v2/caiasset"
 
-	"github.com/GoogleCloudPlatform/terraform-google-conversion/v2/cai2hcl/converters/google/common"
+	"github.com/GoogleCloudPlatform/terraform-google-conversion/v2/cai2hcl/common"
 	tfschema "github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/zclconf/go-cty/cty"
 	"google.golang.org/api/compute/v1"
