+++ cai2hcl/services/resourcemanager/project.go
@@ -1,11 +1,11 @@
-package resources
+package resourcemanager
 
 import (
 	"encoding/json"
 	"fmt"
 	"strings"
 
-	"github.com/GoogleCloudPlatform/terraform-google-conversion/v2/cai2hcl/converters/google/common"
+	"github.com/GoogleCloudPlatform/terraform-google-conversion/v2/cai2hcl/common"
 	"github.com/GoogleCloudPlatform/terraform-google-conversion/v2/caiasset"
 
 	tfschema "github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
